Factor out part of StraightenWidget into base cls
The logic of keybindings, entering and leaving should be the same for
all complex transform widgets and can be re-used for future
implementations such as cropping.
